Refactor files paths
