+++ src/components/Header.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import { Link } from "react-router-dom";
 
-import { isAuthenticated, logoutUser, isAdmin } from "../apis/auth";
+import { isAuthenticated, logoutUser, isAdmin } from "../services/auth";
 
 class Header extends Component {
   render() {
