+++ src/constants/jwt.js
@@ -1,0 +1,3 @@
+export const TOKEN_KEY = "@berlim-token";
+export const USER_TYPE_KEY = "@berlim-user-type";
+export const USER_ID_KEY = "@berlim-user-id";
